refactor: initial draft of mega-refactor
