+++ tb_discord/tb_embeds/__init__.py
@@ -1,0 +1,3 @@
+"""Discord embed subclasses for self-containted updates"""
+import tb_discord.tb_embeds.private
+import tb_discord.tb_embeds.public
