+++ tb_discord/tb_embeds/public.py
@@ -1,0 +1,6 @@
+from discord import Embed as _Embed
+
+
+class EventEmbed(_Embed):
+	def __init(self):
+		super().__init__()
