+++ tb_discord/tb_events.py
@@ -1,0 +1,54 @@
+from datetime import datetime
+from discord import File
+from os import remove
+from PIL import Image, ImageDraw, ImageFont
+from tb_discord import bot
+import random
+import re
+
+
+__all__ = []
+
+
+JETS = ["F16", "F18", "F15", "F35", "F22", "A10", "F14", "MIR2"]
+HOLDING_POINTS = ["A", "B", "C", "D"]
+AERODROMES = ["UG5X", "UG24", "UGKO", "UGKS", "URKA", "URKN", "URMM", "URSS"]
+RUNWAYS = ["22", "04"]
+DEPARTURES = ["GAM1D", "PAL1D", "ARN1D", "TIB1D", "SOR1D", "RUD1D", "AGI1D", "DIB1D", "TUN1D", "NAL1D"]
+
+
+@bot.event
+async def on_ready():
+    print(f"{bot.user} has connected to Discord!")
+
+
+@bot.event
+async def on_member_join(member):
+    time = str(datetime.now())
+    strip_text = {
+        "slot": f"{time[11:16]}",
+        "squawk": f"{random.randint(0, 6)}{random.randint(0, 7)}{random.randint(0, 7)}{random.randint(0, 7)}",
+        "callsign": (f"{re.sub('[^a-zA-Z0-9]', '', member.name.upper())[:4]}" 
+                     f"{random.randint(0, 9)}{random.randint(0, 9)}"),
+        "aircraft": f"{random.choice(JETS)}",
+        "hold": f"{random.choice(HOLDING_POINTS)}",
+        "aerodrome": f"{random.choice(AERODROMES)}",
+        "runway": f"{random.choice(RUNWAYS)}",
+        "departure": f"{random.choice(DEPARTURES)}"
+    }
+    strip = Image.open("strip_blank.png")
+    font = ImageFont.truetype("consolas.ttf", 40)
+    font_large = ImageFont.truetype("consolas.ttf", 70)
+    d = ImageDraw.Draw(strip)
+    d.text((67, 101), strip_text["slot"], font=font, fill=(0, 0, 0), anchor="mm")
+    d.text((305, 101), strip_text["callsign"], font=font_large, fill=(0, 0, 0), anchor="lm")
+    d.text((914, 101), strip_text["hold"], font=font_large, fill=(0, 0, 0), anchor="mm")
+    d.text((1339, 150), strip_text["aerodrome"], font=font, fill=(0, 0, 0), anchor="mm")
+    d.text((1632, 101), strip_text["runway"], font=font_large, fill=(0, 0, 0), anchor="mm")
+    d.text((1803, 101), strip_text["departure"], font=font_large, fill=(0, 0, 0), anchor="mm")
+    d.text((646, 150), strip_text["squawk"], font=font, fill=(0, 0, 0), anchor="lm")
+    d.text((646, 57), "M/" + strip_text["aircraft"], font=font, fill=(0, 0, 0), anchor="lm")
+    strip.save(fp="strip.png")
+    await bot.get_channel(1099805424934469652).send(f"Welcome to Digital Controllers, "
+                                                    f"{member.name}!", file=File("strip.png"))
+    remove("strip.png")
