--- tb_embeds/__init__.py
@@ -1,3 +1,0 @@
-"""Discord embed subclasses for self-containted updates"""
-import private
-import public
